# apps/galley/internal/httpapi/api.gen.go
//go:build go1.22

// Package httpapi provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package httpapi

import (
	"errors"
	"fmt"
	"net/http"

	"github.com/oapi-codegen/runtime"
)

// Defines values for DatabaseStatusStatus.
const (
	DatabaseStatusStatusError DatabaseStatusStatus = "error"
	DatabaseStatusStatusOk    DatabaseStatusStatus = "ok"
)

// Valid indicates whether the value is a known member of the DatabaseStatusStatus enum.
func (e DatabaseStatusStatus) Valid() bool {
	switch e {
	case DatabaseStatusStatusError:
		return true
	case DatabaseStatusStatusOk:
		return true
	default:
		return false
	}
}

// Defines values for StatusResponseApplication.
const (
	Galley StatusResponseApplication = "galley"
)

// Valid indicates whether the value is a known member of the StatusResponseApplication enum.
func (e StatusResponseApplication) Valid() bool {
	switch e {
	case Galley:
		return true
	default:
		return false
	}
}

// Defines values for StatusResponseEnvironment.
const (
	Development StatusResponseEnvironment = "development"
	Production  StatusResponseEnvironment = "production"
)

// Valid indicates whether the value is a known member of the StatusResponseEnvironment enum.
func (e StatusResponseEnvironment) Valid() bool {
	switch e {
	case Development:
		return true
	case Production:
		return true
	default:
		return false
	}
}

// Defines values for StatusResponseStatus.
const (
	StatusResponseStatusOk StatusResponseStatus = "ok"
)

// Valid indicates whether the value is a known member of the StatusResponseStatus enum.
func (e StatusResponseStatus) Valid() bool {
	switch e {
	case StatusResponseStatusOk:
		return true
	default:
		return false
	}
}

// Defines values for TicketStatus.
const (
	Backlog TicketStatus = "Backlog"
)

// Valid indicates whether the value is a known member of the TicketStatus enum.
func (e TicketStatus) Valid() bool {
	switch e {
	case Backlog:
		return true
	default:
		return false
	}
}

// CreateDiagnosticNoteRequest defines model for CreateDiagnosticNoteRequest.
type CreateDiagnosticNoteRequest struct {
	Note string `json:"note"`
}

// CreateTicketRequest defines model for CreateTicketRequest.
type CreateTicketRequest struct {
	// Title Trimmed of leading/trailing whitespace before validation. Must be non-empty and at most 200 characters after trimming.
	Title string `json:"title"`
}

// DatabaseStatus Added by issue #52. Reachability and applied migration version, checked live on every request -- never cached from process start, and never reported as "ok" when the database is unreachable.
type DatabaseStatus struct {
	// Error Present only when status is "error". A short, fixed, non-secret explanation -- never the underlying driver error text or any part of the connection string.
	Error *string `json:"error,omitempty"`

	// MigrationVersion The most recently applied forward-only migration version (see apps/galley/internal/migrations), or null if it could not be determined -- either because the database is unreachable (status "error") or because no migration has been applied yet on an otherwise-reachable database (status "ok").
	MigrationVersion *int                 `json:"migrationVersion"`
	Status           DatabaseStatusStatus `json:"status"`
}

// DatabaseStatusStatus defines model for DatabaseStatus.Status.
type DatabaseStatusStatus string

// DiagnosticNote Development-only diagnostic record (issue #52), persisted in the diagnostic_notes table. Not a domain/ticket concept.
type DiagnosticNote struct {
	// CreatedAt RFC3339 UTC timestamp of when the note was persisted.
	CreatedAt string `json:"createdAt"`
	Id        int    `json:"id"`
	Note      string `json:"note"`
}

// DiagnosticNoteList defines model for DiagnosticNoteList.
type DiagnosticNoteList struct {
	Notes []DiagnosticNote `json:"notes"`
}

// ErrorBody The shared JSON error shape used by every Galley error response.
type ErrorBody struct {
	Error ErrorDetail `json:"error"`
}

// ErrorDetail defines model for ErrorDetail.
type ErrorDetail struct {
	// Code Short, stable, snake_case identifier (e.g. "not_found").
	Code string `json:"code"`

	// Message Human-readable, non-secret explanation.
	Message string `json:"message"`
}

// Owner ticketIt's stable internal Owner identity -- independent of any GitHub identifier (docs/deployment.md, "Ownership and sign-in"). `login` is the linked GitHub identity's most recently observed login, shown for display only: matching a sign-in to this Owner always uses the immutable provider account id, never this field.
type Owner struct {
	// Id ticketIt's own Owner id -- not a GitHub identifier.
	Id int `json:"id"`

	// Login The linked GitHub account's current login, for display only.
	Login string `json:"login"`
}

// SessionResponse defines model for SessionResponse.
type SessionResponse struct {
	// Owner ticketIt's stable internal Owner identity -- independent of any GitHub identifier (docs/deployment.md, "Ownership and sign-in"). `login` is the linked GitHub identity's most recently observed login, shown for display only: matching a sign-in to this Owner always uses the immutable provider account id, never this field.
	Owner Owner `json:"owner"`
}

// StatusResponse The fixed GET /api/status payload.
type StatusResponse struct {
	Application StatusResponseApplication `json:"application"`

	// Database Added by issue #52. Reachability and applied migration version, checked live on every request -- never cached from process start, and never reported as "ok" when the database is unreachable.
	Database    DatabaseStatus            `json:"database"`
	Environment StatusResponseEnvironment `json:"environment"`

	// StartedAt RFC3339 UTC timestamp, captured once at process start.
	StartedAt string               `json:"startedAt"`
	Status    StatusResponseStatus `json:"status"`

	// Version Galley's configured version string (GALLEY_VERSION).
	Version string `json:"version"`
}

// StatusResponseApplication defines model for StatusResponse.Application.
type StatusResponseApplication string

// StatusResponseEnvironment defines model for StatusResponse.Environment.
type StatusResponseEnvironment string

// StatusResponseStatus defines model for StatusResponse.Status.
type StatusResponseStatus string

// Ticket ticketIt's first domain record (issue #56): a title captured in Backlog. No work-type/category column -- see docs/ticket-creation.md, "Flexible ticket structure". Owned by exactly one Owner, enforced by Galley (docs/adr/0001-single-authority-galley.md). Addressed by an opaque, non-sequential public identifier (issue #57) -- see `id` below.
type Ticket struct {
	// CreatedAt RFC3339 UTC timestamp of when the Ticket was captured.
	CreatedAt string `json:"createdAt"`

	// Id Opaque public identifier (issue #57), used in URLs and by GET /api/tickets/{id}. Non-sequential and non-guessable -- never the internal sequential database id, which no Galley endpoint exposes.
	Id string `json:"id"`

	// Status The Ticket's lifecycle stage (CONTEXT.md, "Status"). This slice only ever produces Backlog -- Ready/In Progress/In Review/Done/Blocked arrive with #60's transitions.
	Status TicketStatus `json:"status"`
	Title  string       `json:"title"`

	// UpdatedAt RFC3339 UTC timestamp of the Ticket's last change. Equal to createdAt until #60 adds transitions.
	UpdatedAt string `json:"updatedAt"`
}

// TicketStatus The Ticket's lifecycle stage (CONTEXT.md, "Status"). This slice only ever produces Backlog -- Ready/In Progress/In Review/Done/Blocked arrive with #60's transitions.
type TicketStatus string

// TicketList The signed-in Owner's Tickets, newest first (createdAt descending, id descending as the tiebreak).
type TicketList struct {
	Tickets []Ticket `json:"tickets"`
}

// CompleteGithubOAuthParams defines parameters for CompleteGithubOAuth.
type CompleteGithubOAuthParams struct {
	Code  *string `form:"code,omitempty" json:"code,omitempty"`
	State *string `form:"state,omitempty" json:"state,omitempty"`

	// Error Present when the provider itself reports failure (e.g. the user denied authorization).
	Error *string `form:"error,omitempty" json:"error,omitempty"`
}

// CreateDiagnosticNoteJSONRequestBody defines body for CreateDiagnosticNote for application/json ContentType.
type CreateDiagnosticNoteJSONRequestBody = CreateDiagnosticNoteRequest

// CreateTicketJSONRequestBody defines body for CreateTicket for application/json ContentType.
type CreateTicketJSONRequestBody = CreateTicketRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// CompleteGithubOAuth Complete GitHub OAuth sign-in
	// (GET /api/auth/github/callback)
	CompleteGithubOAuth(w http.ResponseWriter, r *http.Request, params CompleteGithubOAuthParams)
	// StartGithubOAuth Begin GitHub OAuth sign-in
	// (GET /api/auth/github/start)
	StartGithubOAuth(w http.ResponseWriter, r *http.Request)
	// ListDiagnosticNotes List development diagnostic notes
	// (GET /api/dev/diagnostic-notes)
	ListDiagnosticNotes(w http.ResponseWriter, r *http.Request)
	// CreateDiagnosticNote Persist a development diagnostic note
	// (POST /api/dev/diagnostic-notes)
	CreateDiagnosticNote(w http.ResponseWriter, r *http.Request)
	// SignOut Sign out
	// (DELETE /api/session)
	SignOut(w http.ResponseWriter, r *http.Request)
	// GetSession The signed-in Owner
	// (GET /api/session)
	GetSession(w http.ResponseWriter, r *http.Request)
	// GetStatus Application status
	// (GET /api/status)
	GetStatus(w http.ResponseWriter, r *http.Request)
	// ListTickets List the signed-in Owner's Tickets
	// (GET /api/tickets)
	ListTickets(w http.ResponseWriter, r *http.Request)
	// CreateTicket Capture a Ticket from a title alone
	// (POST /api/tickets)
	CreateTicket(w http.ResponseWriter, r *http.Request)
	// GetTicket Get one Ticket by its identifier
	// (GET /api/tickets/{id})
	GetTicket(w http.ResponseWriter, r *http.Request, id string)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// CompleteGithubOAuth operation middleware
func (siw *ServerInterfaceWrapper) CompleteGithubOAuth(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// Parameter object where we will unmarshal all parameters from the context
	var params CompleteGithubOAuthParams

	// ------------- Optional query parameter "code" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "code", r.URL.Query(), &params.Code, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "code"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "code", Err: err})
		}
		return
	}

	// ------------- Optional query parameter "state" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "state", r.URL.Query(), &params.State, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "state"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "state", Err: err})
		}
		return
	}

	// ------------- Optional query parameter "error" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "error", r.URL.Query(), &params.Error, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "error"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "error", Err: err})
		}
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CompleteGithubOAuth(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// StartGithubOAuth operation middleware
func (siw *ServerInterfaceWrapper) StartGithubOAuth(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.StartGithubOAuth(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ListDiagnosticNotes operation middleware
func (siw *ServerInterfaceWrapper) ListDiagnosticNotes(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListDiagnosticNotes(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CreateDiagnosticNote operation middleware
func (siw *ServerInterfaceWrapper) CreateDiagnosticNote(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateDiagnosticNote(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// SignOut operation middleware
func (siw *ServerInterfaceWrapper) SignOut(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.SignOut(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetSession operation middleware
func (siw *ServerInterfaceWrapper) GetSession(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetSession(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetStatus operation middleware
func (siw *ServerInterfaceWrapper) GetStatus(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetStatus(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ListTickets operation middleware
func (siw *ServerInterfaceWrapper) ListTickets(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListTickets(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CreateTicket operation middleware
func (siw *ServerInterfaceWrapper) CreateTicket(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateTicket(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetTicket operation middleware
func (siw *ServerInterfaceWrapper) GetTicket(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", r.PathValue("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetTicket(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of [http.ServeMux].
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	http.Handler
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/api/status", wrapper.GetStatus)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/api/dev/diagnostic-notes", wrapper.ListDiagnosticNotes)
	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/api/dev/diagnostic-notes", wrapper.CreateDiagnosticNote)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/api/auth/github/start", wrapper.StartGithubOAuth)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/api/auth/github/callback", wrapper.CompleteGithubOAuth)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/api/tickets", wrapper.ListTickets)
	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/api/tickets", wrapper.CreateTicket)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/api/tickets/{id}", wrapper.GetTicket)
	m.HandleFunc(http.MethodDelete+" "+options.BaseURL+"/api/session", wrapper.SignOut)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/api/session", wrapper.GetSession)

	return m
}