# contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: ticketIt Galley API
  version: "0.1.0"
  description: >-
    Galley is the sole authority over ticketIt's authoritative records
    (docs/adr/0001-single-authority-galley.md). This document covers the
    subset of its HTTP API that has shipped so far.
servers:
  - url: /
    description: Same-origin — Galley directly, or Swiftlet's dev-server proxy for /api/*.
paths:
  /api/status:
    get:
      operationId: getStatus
      summary: Application status
      description: >-
        Fixed, unauthenticated status payload. `application` and `status`
        are constant; `version` and `environment` come from Galley's
        configuration; `startedAt` is the process start time, captured
        once. Additive-only: fields are never renamed or removed.
        `database` was added by issue #52 and reflects a live check made
        while handling this request, not a cached boot-time snapshot.
      responses:
        "200":
          description: Galley is running.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        default:
          description: >-
            Error. Every Galley error response uses this shape, including
            ones with no operation of their own (unmatched route,
            disallowed method).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/dev/diagnostic-notes:
    get:
      operationId: listDiagnosticNotes
      summary: List development diagnostic notes
      tags: [dev-diagnostic]
      description: >-
        Development-only diagnostic (issue #52): lists every persisted
        diagnostic note, oldest first. Proves data survives a Galley
        restart against the same database. These routes are gated at
        route registration and do not exist when `GALLEY_ENVIRONMENT`
        is `production` — an unmatched request there returns the same
        `404 not_found` shape as any other unknown route.
      responses:
        "200":
          description: The persisted diagnostic notes, oldest first.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiagnosticNoteList"
        default:
          description: Error. See `ErrorBody`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
    post:
      operationId: createDiagnosticNote
      summary: Persist a development diagnostic note
      tags: [dev-diagnostic]
      description: >-
        Development-only diagnostic (issue #52): persists one note.
        Gated the same way as the GET above.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDiagnosticNoteRequest"
      responses:
        "201":
          description: The persisted diagnostic note.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiagnosticNote"
        default:
          description: Error. See `ErrorBody`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/auth/github/start:
    get:
      operationId: startGithubOAuth
      summary: Begin GitHub OAuth sign-in
      tags: [auth]
      description: >-
        Redirects the browser to the configured GitHub OAuth provider's
        authorize endpoint (`GALLEY_OAUTH_GITHUB_BASE_URL`). Issues a
        fresh, high-entropy `state`, persists a hash of it with a short
        expiry, and binds it to this browser via a short-lived,
        HttpOnly `state` cookie -- the callback rejects any request
        whose `state` query parameter does not match both that cookie
        and the persisted record, which is what defeats replay and
        cross-session use. Unauthenticated by design: this is how a
        session is obtained in the first place.
      responses:
        "302":
          description: >-
            Redirect to the provider's authorize endpoint. Sets the
            `state` cookie.
          headers:
            Location:
              schema: {type: string}
        default:
          description: Error. See `ErrorBody`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/auth/github/callback:
    get:
      operationId: completeGithubOAuth
      summary: Complete GitHub OAuth sign-in
      tags: [auth]
      description: >-
        Validates `state` against the cookie set by the start endpoint
        and the persisted record -- rejecting a missing, mismatched,
        expired, or already-used value with `invalid_oauth_state` --
        then exchanges `code` for an access token, fetches the identity
        from the provider, and compares it to the configured Owner. A
        non-owner identity is rejected with `owner_mismatch`: no
        session, Owner record, or link is created or modified. The
        access token is discarded immediately after fetching the
        identity: never stored, never reused for any provider API call
        (see docs/evidence/m2/54-oauth-session.md). On success, creates
        or verifies the linked Owner identity -- matching the immutable
        provider account id, not the login -- persists a new session,
        sets the session cookie, and redirects to the application root.
      parameters:
        - name: code
          in: query
          required: false
          schema: {type: string}
        - name: state
          in: query
          required: false
          schema: {type: string}
        - name: error
          in: query
          required: false
          description: >-
            Present when the provider itself reports failure (e.g. the
            user denied authorization).
          schema: {type: string}
      responses:
        "302":
          description: >-
            Sign-in succeeded. Sets the session cookie and redirects to
            the application root.
          headers:
            Location:
              schema: {type: string}
        default:
          description: >-
            Error. See `ErrorBody`. Includes `invalid_oauth_state`,
            `owner_mismatch`, and `oauth_provider_error` -- see
            apps/galley/README.md, "Error shape."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/tickets:
    get:
      operationId: listTickets
      summary: List the signed-in Owner's Tickets
      tags: [tickets]
      description: >-
        Returns every Ticket belonging to the signed-in Owner, newest
        first: ordered by createdAt descending with id descending as
        the deterministic tiebreak (createdAt alone is not unique --
        see apps/galley/README.md, "Ticket ordering"). Requires a
        valid session; returns `401 unauthenticated` otherwise.
      responses:
        "200":
          description: The Owner's Tickets, newest first.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketList"
        default:
          description: Error. See `ErrorBody`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
    post:
      operationId: createTicket
      summary: Capture a Ticket from a title alone
      tags: [tickets]
      description: >-
        Creates a Ticket owned by the signed-in Owner, in Backlog
        (issue #56: title-only quick capture --
        docs/ticket-creation.md, "Quick capture"). `title` is required,
        is trimmed of leading/trailing whitespace, and must be
        non-empty and at most 200 characters after trimming;
        violations return `invalid_request`. No work-type or category
        is accepted or stored -- Tickets stay generic
        (docs/ticket-creation.md, "Flexible ticket structure"). Requires
        a valid session; returns `401 unauthenticated` otherwise.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTicketRequest"
      responses:
        "201":
          description: The persisted Ticket, in Backlog.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        default:
          description: Error. See `ErrorBody`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/tickets/{id}:
    get:
      operationId: getTicket
      summary: Get one Ticket by its identifier
      tags: [tickets]
      description: >-
        Returns one Ticket belonging to the signed-in Owner, addressed
        by its opaque public identifier (issue #57) -- never the
        internal sequential database id, which no Galley endpoint
        exposes. An unknown identifier, a malformed identifier, and an
        identifier belonging to another Owner all return the same
        404 not_found -- this endpoint never reveals that a record
        exists but belongs to someone else. Requires a valid session;
        returns 401 unauthenticated otherwise.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            # Kept a plain Go string (see Ticket.id's own x-go-type):
            # the handler validates and rejects a malformed value
            # itself, folding it into the same 404 as "not found"
            # rather than a generated binding error in a different
            # shape. See apps/galley/README.md, "Tickets", "Public
            # identifier".
            x-go-type: string
      responses:
        "200":
          description: The Ticket.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        default:
          description: Error. See `ErrorBody`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/session:
    get:
      operationId: getSession
      summary: The signed-in Owner
      tags: [auth]
      description: >-
        Requires a valid session; returns `401 unauthenticated`
        otherwise, in the shared error shape. Unlike `/api/status`,
        this route is never public.
      responses:
        "200":
          description: A session is active.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        default:
          description: Error. See `ErrorBody`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
    delete:
      operationId: signOut
      summary: Sign out
      tags: [auth]
      description: >-
        Revokes the current session (deletes its persisted record) and
        clears the session cookie. Requires a valid session; returns
        `401 unauthenticated` otherwise.
      responses:
        "204":
          description: Session revoked.
        default:
          description: Error. See `ErrorBody`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
components:
  schemas:
...
    SessionResponse:
      type: object
      additionalProperties: false
      required: [owner]
      properties:
        owner:
          $ref: "#/components/schemas/Owner"
...
    TicketList:
      type: object
      description: >-
        The signed-in Owner's Tickets, newest first (createdAt
        descending, id descending as the tiebreak).
      additionalProperties: false
      required: [tickets]
      properties:
        tickets:
          type: array
          items:
            $ref: "#/components/schemas/Ticket"
    CreateTicketRequest:
      type: object
      additionalProperties: false
      required: [title]
      properties:
        title:
          type: string
          description: >-
            Trimmed of leading/trailing whitespace before validation.
            Must be non-empty and at most 200 characters after
            trimming.
          minLength: 1
          maxLength: 200
    StatusResponse:
      type: object
      description: The fixed GET /api/status payload.
      additionalProperties: false
      required: [application, status, version, environment, startedAt, database]
      properties:
        application:
          type: string
          const: galley
        status:
          type: string
          const: ok
        version:
          type: string
          description: Galley's configured version string (GALLEY_VERSION).
        environment:
          type: string
          enum: [development, production]
        startedAt:
          type: string
          format: date-time
          description: RFC3339 UTC timestamp, captured once at process start.
          # Keeps this a Go string. Left alone, date-time maps to
          # time.Time, whose RFC3339Nano marshaling can add fractional
          # seconds and change the response bytes.
          x-go-type: string
        database:
          $ref: "#/components/schemas/DatabaseStatus"
...
    DiagnosticNote:
      type: object
      description: >-
        Development-only diagnostic record (issue #52), persisted in
        the diagnostic_notes table. Not a domain/ticket concept.
      additionalProperties: false
      required: [id, note, createdAt]
      properties:
        id:
          type: integer
        note:
          type: string
        createdAt:
          type: string
          format: date-time
          description: RFC3339 UTC timestamp of when the note was persisted.
          x-go-type: string
    DiagnosticNoteList:
      type: object
      additionalProperties: false
      required: [notes]
      properties:
        notes:
          type: array
          items:
            $ref: "#/components/schemas/DiagnosticNote"
    CreateDiagnosticNoteRequest:
      type: object
      additionalProperties: false
      required: [note]
      properties:
        note:
          type: string
          minLength: 1
    ErrorBody:
      type: object
      description: The shared JSON error shape used by every Galley error response.
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
    ErrorDetail:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          description: Short, stable, snake_case identifier (e.g. "not_found").
        message:
          type: string
          description: Human-readable, non-secret explanation.